+++ packages/@codotype/cli/next-app/tailwind.config.js
@@ -1,15 +1,4 @@
+const codotypeTailwindConfig = require("@codotype/ui/tailwind.config")
 module.exports = {
-    // purge: [],
-    purge: false,
-    // https://tailwindcss.com/docs/guides/nextjs#configure-tailwind-to-remove-unused-styles-in-production
-    // purge: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
-    // darkMode: false, // or 'media' or 'class'
-    darkMode: "class",
-    theme: {
-        extend: {},
-    },
-    variants: {
-        extend: {},
-    },
-    plugins: [require("@tailwindcss/forms")],
+    ...codotypeTailwindConfig,
 };