+++ packages/@codotype/ui/.storybook/CodotypeTheme.js
@@ -4,13 +4,12 @@ export default create({
     base: 'dark',
     brandTitle: '@codotype/ui',
     brandUrl: 'https://codotype.org',
-    // brandImage: './lockup.png',
-    // brandImage: 'https://res.cloudinary.com/codotype/image/upload/v1560045005/tech-logos/codotype.png',
+    brandImage: './lockup.png',
     appBorderColor: "#222234",
     appBg: "#111827",
     appContentBg: "#111827",
     barBg: "#111827",
     appBorderRadius: "0.25rem",
     inputBorder: "#222234",
-    barSelectedColor: "#dddddd"
+    barSelectedColor: "#dddddd",
 });