+++ packages/@codotype/ui/src/components/AppContainer/component.tsx
@@ -10,9 +10,10 @@ import { SocialMeta } from "../Meta";
 export function AppContainer(props: {
     plugin: PluginMetadata;
     children: React.ReactNode;
+    darkMode: boolean;
+    setDarkMode: (darkMode: boolean) => void;
 }) {
-    const [dark, setDark] = React.useState(false);
-    const { plugin } = props;
+    const { darkMode, setDarkMode, plugin } = props;
 
     return (
         <React.Fragment>
@@ -22,22 +23,19 @@ export function AppContainer(props: {
             </Head>
             <div
                 className={classnames("min-h-screen pb-32", {
-                    dark: dark,
-                    "bg-gray-900": dark,
-                    "bg-light-background": !dark,
+                    dark: darkMode,
+                    "bg-gray-900": darkMode,
+                    "bg-light-background": !darkMode,
                 })}
             >
                 <PluginNavbar
                     plugin={plugin}
-                    darkModeEnabled={dark}
+                    darkModeEnabled={darkMode}
                     toggleDarkMode={() => {
-                        setDark(!dark);
+                        setDarkMode(!darkMode);
                     }}
                 />
-                <div
-                    className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 h-full dark:bg-gray-800 bg-light-background text-gray-700 dark:text-gray-200"
-                    // style={{ backgroundColor: "#f5f6f9" }}>
-                >
+                <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 h-full dark:bg-gray-900 bg-light-background text-gray-700 dark:text-gray-200">
                     <div className="grid grid-cols-1">
                         <div className="col-span-1 mt-5">{props.children}</div>
                     </div>