+++ packages/@codotype/ui/src/components/navbar/DarkModeButton.tsx
@@ -9,7 +9,6 @@ export function DarkModeButton(props: {
     const { darkModeEnabled, toggleDarkMode } = props;
     return (
         <button
-            // className="focus:outline-none"
             className="focus:outline-none border-transparent text-gray-500 hover:border-gray-300 hover:text-gray-700 inline-flex items-center px-1 pt-1 border-b-2 text-sm font-medium select-none"
             onClick={() => {
                 toggleDarkMode();
@@ -48,9 +47,7 @@ export function DarkModeButton(props: {
                     </svg>
                 )}
                 <span className="ml-2 font-light">
-                    {darkModeEnabled
-                        ? "Light Mode"
-                        : "Dark Mode (Experimental)"}
+                    {darkModeEnabled ? "Light Mode" : "Dark Mode"}
                 </span>
             </div>
         </button>